Reliable branch tracking in repos
Change fury repo pull behavior in that way, it keeps both the refspec
and the commit hash in fury config file. This means that the code
contributor can run `fury repo pull` and push new layer.fury and
another user who downloads it will not get the code from refspec
(that might be updated since that time) but from the exact commit hash.

Task #62