Add the ability to update repositories
This turned into a much larger commit than I had planned. I ended up
removing a lot of implicit paramters, and started work on improving the
lens definitions in light of a bug I discovered with updating schemas.

So most parameters are now explicit, and we can shortly get rid of most
of the `lenses.scala` file.