The layer select command goes crazy when the layer name is wrong #782 V2
V1 - previous related PR
Create a function to verify if a layer exists.
New exception LayersFailure that is thrown when the given name doesn't exist.
Change the Root ImportPath from ImportPath("") to ImportPath("/")

V2 - new version of fix

now it handles '..'

If we are in /platform and we select '..' we return to /.
If we select 'platform',  it get converted to '/platform'.

ref: layerSelectCommandV2