Some simplification of the front-end
Focussing initially on the resources submenu, as a trial for this
simplified way of describing the front-end.

At the moment, everything looks very Haskell-like. There's some
flexibility as to how much we write in for-comprehension style, and how
much we implement using the `>>` and `>>=` operator style.

I still need to provide better support for required options.